requirements.txt
fastapi
uvicorn
supabase
openai
anthropic
mistralai
cohere
google-generativeai
python-dotenv
pydantic
cryptography
